Removed the dependancy on the _desktop_ module.
Users will have to cut-and-paste the auth URL for themselves, but
that is easier then having to figure out that the package is
missing a module.  I did glance at the license and don't think it
would be an issue to include the desktop module's code in the
package, but there is another [issue][profiles] that stems from
the use of this module.  Let's see how things go with out it.

I suspect that the Auth URL will not be able to be selected for a
copy-and-paste action.  Even still, I think there is a better way
to do what needs to be done here.

[profile]: https://github.com/revolunet/sublimetext-google-apps-scripts/issues/4